Move clang flags handling to new `compile.py` file
This mirrors the existing `link.py` file.

My plan is to make use of these functions to call clang directly when
building the system libs.
